refactor: Article, ArticleDetail, ArticleGithubUrl, ArticleTitleUrl 스타일 분리
