+++ src/components/article/Article.tsx
@@ -1,29 +1,11 @@
 /** @jsxImportSource @emotion/react */
-import { css } from '@emotion/react';
 import ArticleDetail from './ArticleDetail';
 import { useEffect, useState } from 'react';
 import SkillsTab from 'components/tabs/SkillsTab';
 import ArticleGithubUrl from './ArticleGithubUrl';
 import { TabPanelProps } from 'components/tabs/SkillsTab';
 import ArticleTitleUrl from './ArticleTitleUrl';
-
-const articleStyle = css({
-    display: 'flex',
-    flexDirection: 'column',
-    rowGap: '25px',
-    marginBottom: '5rem',
-
-    span: {
-        lineHeight: '2.5rem',
-    },
-});
-
-const spanRowStyle = css({
-    display: 'flex',
-    flexDirection: 'row',
-    gap: '1rem',
-    alignItems: 'center',
-});
+import { articleStyle, spanRowStyle } from './Style';
 
 export interface ArticlePropsType {
     title: string;
