+++ src/components/article/ArticleDetail.tsx
@@ -2,18 +2,7 @@
 import { css } from '@emotion/react';
 import CircleIcon from '@mui/icons-material/Circle';
 import { HIGHLITGHT } from 'styles/Colors';
-
-const spanRowStyle = css({
-    display: 'flex',
-    flexDirection: 'row',
-    gap: '0.5rem',
-});
-
-const circleStyle = css({
-    fontSize: '1rem',
-    color: `${HIGHLITGHT}`,
-    marginTop: '0.4rem',
-});
+import { articleDetailStyle, circleStyle } from './Style';
 
 interface detailType {
     detail: string;
@@ -21,7 +10,7 @@ interface detailType {
 
 const ArticleDetail = ({ detail }: detailType) => {
     return (
-        <span css={spanRowStyle}>
+        <span css={articleDetailStyle}>
             <CircleIcon fontSize='small' css={circleStyle} />
             <h3>{detail}</h3>
         </span>
