+++ src/components/article/ArticleGithubUrl.tsx
@@ -1,7 +1,7 @@
 /** @jsxImportSource @emotion/react */
 import GitHubIcon from '@mui/icons-material/GitHub';
 import IconButton from '@mui/material/IconButton';
-import { css } from '@emotion/react';
+import { gitUrlStyle } from './Style';
 import { useLinkToPage } from 'hooks/useLinkToPage';
 import { useEffect } from 'react';
 
@@ -13,14 +13,11 @@ interface githubUrlType {
 const ArticleGithubUrl = ({ githubUrl, title }: githubUrlType) => {
     const buttonTitle = title + 'github';
 
-    const urlStyle = css({
-        lineHeight: '3.7rem',
-    });
     useEffect(() => {
         useLinkToPage(buttonTitle, githubUrl);
     }, [buttonTitle, githubUrl]);
     return (
-        <div id={buttonTitle} css={urlStyle}>
+        <div id={buttonTitle} css={gitUrlStyle}>
             <IconButton>
                 <GitHubIcon />
             </IconButton>
